Remove useless call to prefix_list_cmp() in md_router_cmp()
md_router_cmp() performs a call to prefix_list_cmp() between
the old and new routers to compare their list of avertised
prefixes. The return value (result of the comparison) is not
checked which makes the call useless.

AFAICT, it is not reliable to include a comparison of advertised
prefixes to decide if the two given routers are the same device.
For that reason, the call to prefix_list_cmp() is simply removed.