Set /proc proxy_ndp entry for HA interfaces
When configured as a HA, UMIP needs to act as a ND proxy for its MN
on their home subnet interfaces. At the moment, for that to work,
the administrator needs to explicitly set proxy_ndp key in /proc
for associated interfaces.

The patch below automatically sets proxy_ndp key for HA interfaces
referenced in configuration file. It also unset this key for those
interfaces when leaving. This avoids the need to set/unset those
valid manually.