Directly pass hai to mn_get_ro_coa() instead of hoa.
mn_get_ro_coa() takes the HoA as second parameter and performs
a lookup to get associated home_addr_info structure (hai). The
two  callers of the function have the hai already available. The
patch below changes the prototype of the function to have a
pointer to a home_addr_info structure passed instead of the
HoA.