Simplify ndisc_send_rs() and ndisc_send_ns()
ndisc_send_rs() and ndisc_send_ns() both support two kinds of
calls: either with the unspecified address as source or with
a real address. The latter case is handled via calls to
icmp6_create()/icmp6_send() and the second via a call to
and icmp6_send_unspec().

In practice, the functions are *always* called in UMIP with
the unspecified address as source. This patch takes that fact
into account and simplifies both functions in order to support
only the useful case where the source address is ::.