Gracefully handle multiple interface definition in configuration
When you have an interface referenced multiple times in the HA configuration
file, UMIP just fails to start: the interface index being referenced multiple
times in ha_interface list, this leads UMIP to try and register the all nodes
multicast address (in ha_addr_setup()) twice, which leads to the failure.

On HA with multiple MN, when you handle the configuration of each node in a
separate block or file (a pending patch allows splitting configuration in
multiple files), you *really* want to set the Home Subnet interface of the
node in its file. If multiple nodes are on the same subnet, w/o this patch,
you cannot do that.