Extract CN and HA checks of mh_bu_parse()
When acting as HA, the BU handler (ha_recv_bu()) handles both kinds of
BU, i.e. w/ and w/o the H flag set. This handler simply performs a
call to ha_recv_bu_main() which may pass the processing of the BU to
cn_recv_bu() if H flag is not set.

In both functions (cn_recv_bu() and ha_recv_bu_main()), mh_bu_parse()
is called. That function contains both parsing and checking parts which
are specific to HA and CN cases. For instance, the return path on the
HA ends up in the middle of the function (i.e. the end of the function
is dedicated to CN checks).

This patch:
- simplifies mh_bu_parse() in order to keep only the code common to
  all cases, i.e. the parsing code.
- extracts HA and CN specific checks in two helpers:
  mh_ha_home_bu_check() and mh_cn_bu_check().
- change ha_recv_bu() (the main MH BU handler on the HA) to perform
  the dispatch between HA and CN processing earlier.
- renamed ha_recv_bu_main() to ha_recv_home_bu()
- perform the required change in vt.c where ha_recv_home_bu() is also
  called (i.e. now do CN/HA dispatch)
- move HA and CN specific code to ha.c and cn.c