Make ndisc_send_unspec() more generic
ndisc_send_unspec() is supposed to be generic wrt the type of ND
message it sends using the unspecified address as source. At the
moment, it contains NS and RS specific code to fill ICMPv6 header
parameters.

That patch changes the prototype of the function by allowing the
(higher level) caller (ndisc_send_rs() or ndisc_send_ns()) to build
and pass the ICMPv6 header to be used. For obvious reasons, checksum
is still computed in ndisc_send_unspec().