Handle packing issue with struct nd_opt_hdr
In glibc, structures are not packed. This is usually ok but on some
architectures (ARM for instance), this creates issues when checking
the expected length of received packets. For instance, sizeof
returns a size of 4 bytes on ARM architectures for such a structure:

struct nd_opt_hdr      /* Neighbor discovery option header */
  {
    uint8_t  nd_opt_type;
    uint8_t  nd_opt_len;        /* in units of 8 octets */
    /* followed by option specific data */
  };

The patch below basically includes a modified version of glibc's
netinet/icmp6.h in our local copy of the file (include/netinet/icmp6.h)
which adds a __attribute__ ((packed)); to that structure to force packing
in all cases.

The issue was reported on the list by Jozsef Kovacs in November 2008
(http://www.linux-ipv6.org/ml/usagi-users/msg04078.html) and was
previously fixed by non_packed_structures.patch in our patches repo.