Simplify process_link()
In the middle of process_link() a check was done for AF_INET6
ifaces on the length of the L2 address. Another check was done
for loopback interfaces. We now make a more logical check
in the function by looking if we do have the logic to support
the interface type.

The address length check that was done here is now done when
the address is used, in md_create_inet6_iface(). This also avoid
code duplication.