Fix typo in in6_cksum() preventing source address to be considered
The typo was introduced in 7e3b3b79 (Rewrite and move csum() to util module).

It remained unnoticed probably because in6_cksum() is only used by
ndisc_send_unspec(), which *always* passes the unspecified address as
source. Due to the memset() at the beginning of in6_cksum(), the
address used in the pseudo header was always the unspecified address.

Reported-by: Steve Chen <schen@mvista.com>
Signed-off-by: Arnaud Ebalard <arno@natisbad.org>